Initial commit of an IE11(and earlier) compatibility analyser, currently only identifies SelectedOptions accesses
This would probably benefit from a few more tests and the assembly
version numbers would benefit from being tied into some sort of
automated build process.

The SelectedOptions analyser could very naturally be extended with a
code fix (replace any ".SelectedOptions" acceses with ".Options.Where(o
=> o.Selected")) but I've left that as an exercise for another day (it
would be fairly easy in the simple case but there are numerous edge
cases to look after - eg. ensuring that the System.Linq namespace is
available and ensuring that "o" is not already in use in the current
scope).

Still, hopefully this is a worthwhile commit to get things moving (I
started with an analyzer project template, so some of the code - such as
the DiagnosticVerifier class - comes from there and some is custom code
for the logic that we're interested in).